Share unsafe migration syntax helpers
Copilot-Session: e39a2bb4-8a72-40d6-8a1c-d98b6c346afe
