+++ src/tools/illink/src/ILLink.RoslynAnalyzer/ILLink.RoslynAnalyzer.csproj
@@ -50,10 +50,6 @@
     </PackageReference>
   </ItemGroup>
 
-  <ItemGroup>
-    <InternalsVisibleTo Include="ILLink.CodeFixProvider" />
-  </ItemGroup>
-
   <Import Project="..\ILLink.Shared\ILLink.Shared.projitems" Label="Shared" />
 
 </Project>
